Move CustomProperties
