+++ Assets/Puppet/PuppetAppCenter.cs
@@ -17,6 +17,35 @@ public class PuppetAppCenter : MonoBehaviour
 
     static PuppetAppCenter instance;
 
+    public GameObject CustomProperty;
+    public RectTransform PropertiesList;
+
+    public void AddProperty()
+    {
+        var property = Instantiate(CustomProperty);
+        property.transform.SetParent(PropertiesList, false);
+    }
+
+    public void Send()
+    {
+        AppCenter.SetCustomProperties(GetProperties());
+    }
+
+    private CustomProperties GetProperties()
+    {
+        var properties = PropertiesList.GetComponentsInChildren<PuppetCustomProperty>();
+        if (properties == null || properties.Length == 0)
+        {
+            return null;
+        }
+        var result = new CustomProperties();
+        foreach (var i in properties)
+        {
+            i.Set(result);
+        }
+        return result;
+    }
+
     private void Awake()
     {
         Crashes.ShouldProcessErrorReport = ShouldProcessErrorReportHandler;
