--- Assets/Puppet/PuppetCustomProperties.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 98be6d56d081d3f44b7220a847c4ff3f
-timeCreated: 1501070576
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
